mlx-app-converter/manual-tests/gradio-llm-fixture/requirements.txt
gradio==5.0.0
transformers
torch
python-dotenv
